REFACTOR: linting for 2.0 release (#341)
* API: move simulation base classes to new module facet.simulation.base

* REFACTOR: remove an obsolete 'if' clause

* REFACTOR: declare constant for repetitive string literal

* FIX: change a type hint from List to Sequence

* DOC: remove an obsolete comment

* REFACTOR: declare constants for repetitive string literals

* REFACTOR: clarified comments and variable names

* REFACTOR: de-complexify Sample.__init__

* REFACTOR: remove an unused variable

* REFACTOR: disambiguate a regular expression

* REFACTOR: merge a nested 'if' expression

* REFACTOR: improve comments

* REFACTOR: simplify CSS

* REFACTOR: un-bracket alternatives in regular expression

* TEST: remove unused fixtures from test method arguments

* TEST: remove an obsolete to-do

* TEST: anonymize unused loop variables

* TEST: add constants for recurring exception messages

* TEST: minor code tweaks
