+++ src/facet/inspection/_shap.py
@@ -52,6 +52,13 @@ ShapToDataFrameFunction = Callable[
 
 
 #
+# Constants
+#
+
+ASSERTION__CALCULATOR_IS_FITTED = "calculator is fitted"
+
+
+#
 # Ensure all symbols introduced below are included in __all__
 #
 
@@ -289,8 +296,6 @@ class ShapCalculator(
         return shap_tensors
 
     def _preprocess_features(self, sample: Sample) -> pd.DataFrame:
-        # get the out-of-bag subsample of the training sample, with feature columns
-        # in the sequence that was used to fit the learner
 
         # get the model
         pipeline = self.pipeline
@@ -324,7 +329,7 @@ class ShapCalculator(
         pass
 
     @abstractmethod
-    def _get_output_names(self, sample: Sample) -> List[str]:
+    def _get_output_names(self, sample: Sample) -> Sequence[str]:
         pass
 
 
@@ -367,7 +372,7 @@ class ShapValuesCalculator(
 
         multi_output_type = self.get_multi_output_type()
         multi_output_names = self.get_multi_output_names(sample=sample)
-        assert self.feature_index_ is not None, "Calculator is fitted"
+        assert self.feature_index_ is not None, ASSERTION__CALCULATOR_IS_FITTED
         features_out = self.feature_index_
 
         # calculate the shap values, and ensure the result is a list of arrays
@@ -409,14 +414,14 @@ class ShapInteractionValuesCalculator(
         """[see superclass]"""
 
         self.ensure_fitted()
-        assert self.shap_ is not None, "Calculator is fitted"
+        assert self.shap_ is not None, ASSERTION__CALCULATOR_IS_FITTED
         return self.shap_.groupby(level=0).sum()
 
     def get_shap_interaction_values(self) -> pd.DataFrame:
         """[see superclass]"""
 
         self.ensure_fitted()
-        assert self.shap_ is not None, "Calculator is fitted"
+        assert self.shap_ is not None, ASSERTION__CALCULATOR_IS_FITTED
         return self.shap_
 
     def get_diagonals(self) -> pd.DataFrame:
@@ -435,7 +440,7 @@ class ShapInteractionValuesCalculator(
             self.shap_ is not None
             and self.sample_ is not None
             and self.feature_index_ is not None
-        ), "Calculator is fitted"
+        ), ASSERTION__CALCULATOR_IS_FITTED
 
         n_observations = len(self.sample_)
         n_features = len(self.feature_index_)
@@ -462,7 +467,7 @@ class ShapInteractionValuesCalculator(
 
         multi_output_type = self.get_multi_output_type()
         multi_output_names = self.get_multi_output_names(sample)
-        assert self.feature_index_ is not None, "Calculator is fitted"
+        assert self.feature_index_ is not None, ASSERTION__CALCULATOR_IS_FITTED
         features_out = self.feature_index_
 
         # calculate the shap interaction values; ensure the result is a list of arrays
@@ -601,7 +606,7 @@ class ClassifierShapCalculator(ShapCalculator[ClassifierPipelineDF], metaclass=A
     def _get_output_names(
         self,
         sample: Sample,
-    ) -> List[str]:
+    ) -> Sequence[str]:
         assert not isinstance(
             sample.target_name, list
         ), "classification model is single-output"
