+++ src/facet/simulation/__init__.py
@@ -4,6 +4,7 @@ selected fixed values of a selected input feature.
 
 For further details on the simulation approach see the
 :ref:`Bootstrap simulation with FACET<boot-sim-summary-tut>` tutorial.
-
 """
+
+from ._result import *
 from ._simulation import *
