+++ src/facet/simulation/base/__init__.py
@@ -1,0 +1,4 @@
+"""
+Base classes for simulation.
+"""
+from ._base import *
