+++ test/test/facet/test_inspection.py
@@ -193,7 +193,6 @@ def test_model_inspection_classifier_binary_single_shap_output() -> None:
 # noinspection DuplicatedCode
 def test_model_inspection_classifier_multi_class(
     iris_inspector_multi_class: LearnerInspector[ClassifierPipelineDF],
-    n_jobs: int,
 ) -> None:
     iris_classifier = iris_inspector_multi_class.pipeline
     iris_sample = iris_inspector_multi_class.sample_
@@ -399,7 +398,6 @@ def test_model_inspection_classifier_interaction(
     # shap interaction values add up to shap values
     # we have to live with differences of up to 0.020, given the different results
     # returned for SHAP values and SHAP interaction values
-    # todo: review accuracy after implementing use of a background dataset
     assert (
         model_inspector_no_interaction.shap_values() - shap_values
     ).abs().max().max() < 0.015
