+++ test/test/facet/test_partition.py
@@ -7,11 +7,17 @@ from facet.data.partition import (
     IntegerRangePartitioner,
 )
 
+# recurring exception messages
+MSG_ARG_VALUES_EMPTY = "arg values is empty"
+MSG_INSUFFICIENT_VARIANCE = (
+    "insufficient variance in values; cannot infer partitioning bounds"
+)
+
 
 def test_discrete_partitioning() -> None:
     np.random.seed(42)
 
-    for i in range(10):
+    for _ in range(10):
 
         values = np.random.randint(
             low=0, high=10000, size=np.random.randint(low=100, high=200)
@@ -44,7 +50,7 @@ def test_discrete_partitioning() -> None:
 def test_continuous_partitioning() -> None:
     np.random.seed(42)
 
-    for i in range(10):
+    for _ in range(10):
 
         values = np.random.normal(
             loc=3.0, scale=8.0, size=np.random.randint(low=2000, high=4000)
@@ -79,7 +85,7 @@ def test_continuous_partitioning() -> None:
 
 def test_category_partitioning() -> None:
     np.random.seed(42)
-    for i in range(10):
+    for _ in range(10):
         values = np.random.randint(
             low=0, high=10, size=np.random.randint(low=100, high=200)
         )
@@ -97,43 +103,43 @@ def test_partition_with_invalid_values() -> None:
 
     with pytest.raises(
         ValueError,
-        match="arg values is empty",
+        match=MSG_ARG_VALUES_EMPTY,
     ):
         ContinuousRangePartitioner().fit([])
 
     with pytest.raises(
         ValueError,
-        match="insufficient variance in values; cannot infer partitioning bounds",
+        match=MSG_INSUFFICIENT_VARIANCE,
     ):
         ContinuousRangePartitioner().fit([1])
 
     with pytest.raises(
         ValueError,
-        match="insufficient variance in values; cannot infer partitioning bounds",
+        match=MSG_INSUFFICIENT_VARIANCE,
     ):
         ContinuousRangePartitioner().fit([1, 1, 1, 10, 1])
 
     with pytest.raises(
         ValueError,
-        match="arg values is empty",
+        match=MSG_ARG_VALUES_EMPTY,
     ):
         IntegerRangePartitioner().fit([])
 
     with pytest.raises(
         ValueError,
-        match="insufficient variance in values; cannot infer partitioning bounds",
+        match=MSG_INSUFFICIENT_VARIANCE,
     ):
         IntegerRangePartitioner().fit([1])
 
     with pytest.raises(
         ValueError,
-        match="insufficient variance in values; cannot infer partitioning bounds",
+        match=MSG_INSUFFICIENT_VARIANCE,
     ):
         IntegerRangePartitioner().fit([1, 1, 1, 10, 1])
 
     with pytest.raises(
         ValueError,
-        match="arg values is empty",
+        match=MSG_ARG_VALUES_EMPTY,
     ):
         CategoryPartitioner().fit([])
